bug fix: copy size
This turns out to be more serious than I initially thought. For tree
walk, only rank 0 needs to create() and through the root of tree into
the queue. For copy, since we pre-walk the tree, each task is holding a
copy files they walked. The create() process need to enq() for each
file. That is, every process needs to create(). The change make the
calling for every process "create()" a default. The treewalk is treated
as a special case: within create() of pwalk, it only do enq() if its
rank is 0.