Add checkpoint and restart feature
PCP now will periodically checkpoint the work queue. We don't try to
recover if the application fails during the tree walk stage. The
checkpoint file is named after a restart ID, and this ID is needed if
the restart (prestart) is to be run later. I expect this to be JOB_ID in
a batch sytem. When not provided, a current timestamp is used.