BUG FIX: core circle send_work() dup reply message
This is a sinister bug to fix: it only materialize through repeated runs
on a fairly large test dataset (100k files). The iteration can go as
high as 130 runs without problem. The root cause is when distributing
work loads, some processes could receive zero tasks. In those cases, we
need to send_no_work(); upon sending the no work message, the correct
course of action is to return immediately. If not - that is where the
bug is, the logic continues and it will send a empty list [], instead of
value 0 as it should in the send_no_work() case.