A prototype for embeded rsync
Basic ideal is that when --force and --rsync are provided, we will
follow rsync algorithm and deal with those files already exists. The
non-exist file will still FCP'ed. The librsync should be more efficient,
but its wrapper have one serious bug and I am not sure I want to spend
time wrapping it.