reformat fprof output
by request of dustin, no need to generate visual block diagram. Also, a
perecentage of both file # and file size are good to have on the output.
The workaround is, the total fileset size are collected when we do
"treewalk.epilogue()", so the output from epilogue will be printed
first, and followed by fileset histogram.